Move lookup endpoint to CS API (and s/is_server/id_server/)
