+++ synapse/config/server.py
@@ -573,7 +573,6 @@ KNOWN_RESOURCES = (
     'replication',
     'static',
     'webclient',
-    'identity',
 )
