--- synapse/rest/identity/v1/__init__.py
@@ -1,24 +1,0 @@
-# -*- coding: utf-8 -*-
-# Copyright 2019 New Vector Ltd
-#
-# Licensed under the Apache License, Version 2.0 (the "License");
-# you may not use this file except in compliance with the License.
-# You may obtain a copy of the License at
-#
-#     http://www.apache.org/licenses/LICENSE-2.0
-#
-# Unless required by applicable law or agreed to in writing, software
-# distributed under the License is distributed on an "AS IS" BASIS,
-# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-# See the License for the specific language governing permissions and
-# limitations under the License.
-
-from twisted.web.resource import Resource
-
-from .lookup import IdentityLookup
-
-
-class IdentityApiV1Resource(Resource):
-    def __init__(self, hs):
-        Resource.__init__(self)
-        self.putChild(b"lookup", IdentityLookup(hs))
